change the definition of int64 and uint64 for the compatibility of 32 bit and 64 bit system

+++ model_utils.hpp
@@ -22,16 +22,16 @@ namespace med {
     /**
      *  data type definition
      */
-    typedef char   int8;
-    typedef short  int16;
-    typedef int    int32;
-    typedef long   int64;
-    typedef float  float32;
-    typedef double float64;
-    typedef unsigned char  uint8;
-    typedef unsigned short uint16;
-    typedef unsigned int   uint32;
-    typedef unsigned long  uint64;
+    typedef char        int8;
+    typedef short       int16;
+    typedef int         int32;
+    typedef long long   int64;
+    typedef float       float32;
+    typedef double      float64;
+    typedef unsigned char       uint8;
+    typedef unsigned short      uint16;
+    typedef unsigned int        uint32;
+    typedef unsigned long long  uint64;
     
     /**
      *  utils for reading and writing
